requirements.txt
pip~=24.2
pillow~=11.3.0
redis~=6.4.0
Django~=5.0.3
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
django-cors-headers~=4.3.1
django-environ~=0.11.2

